Don't always rebuild existing image

+++ Makefile
@@ -6,9 +6,8 @@ GDVER:=$(shell grep -oP 'ARG GDVER=\K.*' Dockerfile)
 REL=$(TAG):$(PYVER)-gdal$(GDVER)
 GID=$(shell id -g)
 
-.PHONY: run bash push push-docker push-github tag build force
+.PHONY: run bash push push-docker push-github tag build force ls rm
 
-# TODO: if $(TAG):lastest exsits, skip build, just run/bash
 run: build
 	docker run --rm -it $(TAG):latest
 bash: build
@@ -56,9 +55,22 @@ push-github: build tag
 tag: build
 	docker tag $(TAG):latest $(REL)
 
+image_exists = $(shell docker image inspect $(TAG):latest --format="exists" 2>/dev/null || echo "")
+last_built = $(shell docker image inspect $(TAG):latest --format='{{.Created}}' 2>/dev/null | cut -d'.' -f1)
+
 build:
+ifeq ($(image_exists),)
 	docker build --progress=plain --tag $(TAG):latest . && \
 	  docker run --rm --volume=$$PWD:/mnt --user=$$UID:$(GID) $(TAG):latest sh -c 'cp /*.txt /mnt/'
+else
+	@echo "Not rebuilding $(TAG):latest built $(last_built). If you don't like that, try \`make rm\` to remove it first."
+endif
 force:
 	docker build --progress=plain --tag $(TAG):latest --pull --no-cache . && \
 	  docker run --rm --volume=$$PWD:/mnt --user=$$UID:$(GID) $(TAG):latest sh -c 'cp /*.txt /mnt/'
+
+ls:
+	@docker image ls |grep --color=always -E "$(TAG)|^REPO"
+
+rm:
+	docker image rm $(TAG):latest
